ошибки + pixel perfect
