+++ gulpfile.js
@@ -2,157 +2,165 @@ import gulp from 'gulp';
 import plumber from 'gulp-plumber';
 import sass from 'gulp-dart-sass';
 import postcss from 'gulp-postcss';
+import autoprefixer from 'autoprefixer';
 import csso from 'postcss-csso';
 import rename from 'gulp-rename';
-import autoprefixer from 'autoprefixer';
-import htmlmin from 'gulp-htmlmin';
 import terser from 'gulp-terser';
-import imagemin from 'gulp-imagemin';
-import webp from 'gulp-webp';
+import squoosh from 'gulp-libsquoosh';
 import svgo from 'gulp-svgmin';
 import svgstore from 'gulp-svgstore';
-import { deleteAsync } from 'del';
+import del from 'del';
 import browser from 'browser-sync';
 
 // Styles
 
-export const styles = (done) => {
-  gulp.src('source/sass/style.scss', { sourcemaps: true })
-    .pipe(plumber())
-    .pipe(sass().on('error', sass.logError))
-    .pipe(postcss([
-      autoprefixer(),
-      csso()
-    ]))
-    .pipe(rename('style.min.css'))
-    .pipe(gulp.dest('build/css', { sourcemaps: '.' }))
-    .pipe(browser.stream());
-  done();
+export const styles = () => {
+return gulp.src('source/sass/style.scss', { sourcemaps: true })
+.pipe(plumber())
+.pipe(sass().on('error', sass.logError))
+.pipe(postcss([
+autoprefixer(),
+csso()
+]))
+.pipe(rename('style.min.css'))
+.pipe(gulp.dest('build/css', { sourcemaps: '.' }))
+.pipe(browser.stream());
 }
 
 // HTML
 
-export const html = () => {
-  return gulp.src('source/*.html')
-    .pipe(htmlmin({ collapseWhitespace: true }))
-    .pipe(gulp.dest('build'));
+const html = () => {
+return gulp.src('source/*.html')
+.pipe(gulp.dest('build'));
 }
 
 // Scripts
 
 const scripts = () => {
-  return gulp.src('source/js/*.js')
-    .pipe(terser())
-    .pipe(gulp.dest('build/js'))
+return gulp.src('source/js/*.js')
+.pipe(gulp.dest('build/js'))
+.pipe(browser.stream());
 }
 
 // Images
 
-export const optimizeImages = () => {
-  return gulp.src('source/img/**/*.{jpg,png}')
-    .pipe(imagemin())
-    .pipe(gulp.dest('build/img'))
+const optimizeImages = () => {
+return gulp.src('source/img/**/*.{png,jpg}')
+.pipe(squoosh())
+.pipe(gulp.dest('build/img'))
 }
 
 const copyImages = () => {
-  return gulp.src('source/img/**/*.{jpg,png}')
-    .pipe(gulp.dest('build/img'))
+return gulp.src('source/img/**/*.{png,jpg}')
+.pipe(gulp.dest('build/img'))
 }
 
 // WebP
 
 const createWebp = () => {
-  return gulp.src('source/img/**/*.{jpg,png}')
-    .pipe(webp())
-    .pipe(gulp.dest('build/img'))
+return gulp.src('source/img/**/*.{png,jpg}')
+.pipe(squoosh({
+webp: {}
+}))
+.pipe(gulp.dest('build/img'))
 }
 
 // SVG
 
 const svg = () =>
-  gulp.src(['source/img/*.svg', '!source/img/icons/*.svg'])
-    .pipe(svgo())
-    .pipe(gulp.dest('build/img'));
-
+gulp.src(['source/img/*.svg', '!source/img/icons/*.svg'])
+.pipe(svgo())
+.pipe(gulp.dest('build/img'));
 
 const sprite = () => {
-  return gulp.src('source/img/icons/*.svg')
-    .pipe(svgo())
-    .pipe(svgstore({
-      inlineSvg: true
-    }))
-    .pipe(rename('sprite.svg'))
-    .pipe(gulp.dest('build/img'))
+return gulp.src('source/img/icons/*.svg')
+.pipe(svgo())
+.pipe(svgstore({
+inlineSvg: true
+}))
+.pipe(rename('sprite.svg'))
+.pipe(gulp.dest('build/img'));
 }
 
-//Copy
+// Copy
 
-const copy = () => {
-  return gulp.src([
-    'source/fonts/**/*.{woff2,woff}',
-    'source/*ico',
-  ], {
-    base: 'source'
-  })
-    .pipe(gulp.dest('build'))
+const copy = (done) => {
+gulp.src([
+'source/fonts/**/*.{woff2,woff}',
+'source/*.ico',
+], {
+base: 'source'
+})
+.pipe(gulp.dest('build'))
+done();
 }
 
 // Clean
 
 const clean = () => {
-  return deleteAsync('build');
-}
+return del('build');
+};
 
 // Server
 
 const server = (done) => {
-  browser.init({
-    server: {
-      baseDir: 'build'
-    },
-    cors: true,
-    notify: false,
-    ui: false,
-  });
-  done();
+browser.init({
+server: {
+baseDir: 'build'
+},
+cors: true,
+notify: false,
+ui: false,
+});
+done();
+}
+
+// Reload
+
+const reload = (done) => {
+browser.reload();
+done();
 }
 
 // Watcher
 
 const watcher = () => {
-  gulp.watch('source/sass/**/*.scss', gulp.series(styles));
-  gulp.watch('source/*.html', gulp.series(html));
+gulp.watch('source/sass/**/*.scss', gulp.series(styles));
+gulp.watch('source/js/*.js', gulp.series(scripts));
+gulp.watch('source/*.html', gulp.series(html, reload));
 }
 
 // Build
 
 export const build = gulp.series(
-  clean,
-  copy,
-  optimizeImages,
-  gulp.parallel(
-    styles,
-    html,
-    scripts,
-    svg,
-    sprite
-  ),
+clean,
+copy,
+optimizeImages,
+gulp.parallel(
+styles,
+html,
+scripts,
+svg,
+sprite,
+createWebp
+),
 );
 
+// Default
 
 export default gulp.series(
-  clean,
-  copy,
-  copyImages,
-  gulp.parallel(
-    styles,
-    html,
-    scripts,
-    svg,
-    sprite,
-    createWebp
-  ),
-  gulp.series(
-    server,
-    watcher
-  ));
+clean,
+copy,
+copyImages,
+gulp.parallel(
+styles,
+html,
+scripts,
+svg,
+sprite,
+createWebp
+),
+gulp.series(
+server,
+watcher
+));
