+++ package.json
@@ -1,36 +1,32 @@
 {
-  "name": "cat-energy-1",
   "private": true,
+  "name": "cat-energy-1",
   "description": "Личный проект «Кэт энерджи» от HTML Academy",
   "devDependencies": {
     "@htmlacademy/editorconfig-cli": "1.0.0",
     "autoprefixer": "10.4.0",
-    "browser-sync": "2.27.7",
-    "del": "^7.1.0",
+    "browser-sync": "^2.23.7",
+    "del": "6.0.0",
     "gulp": "4.0.2",
     "gulp-dart-sass": "^1.0.2",
-    "gulp-htmlmin": "^5.0.1",
-    "gulp-imagemin": "^9.0.0",
+    "gulp-libsquoosh": "^1.0.12",
     "gulp-plumber": "1.2.1",
     "gulp-postcss": "9.0.1",
-    "gulp-rename": "^2.0.0",
+    "gulp-rename": "2.0.0",
     "gulp-svgmin": "^4.1.0",
-    "gulp-svgstore": "^9.0.0",
-    "gulp-terser": "^2.1.0",
-    "gulp-webp": "^5.0.0",
+    "gulp-svgstore": "8.0.0",
+    "gulp-terser": "2.1.0",
     "postcss": "8.3.11",
-    "postcss-csso": "^6.0.1",
+    "postcss-csso": "5.0.1",
     "stylelint": "13.13.1",
-    "stylelint-config-htmlacademy": "0.1.11"
+    "stylelint-config-htmlacademy": "^0.1.11"
   },
   "scripts": {
     "editorconfig": "editorconfig-cli",
     "stylelint": "stylelint \"source/sass/**/*.scss\" --syntax scss",
     "lint": "npm run editorconfig && npm run stylelint",
     "build": "gulp build",
-    "start": "gulp",
-    "tasks": "gulp --tasks",
-    "images": "gulp optimizeImages"
+    "start": "gulp"
   },
   "browserslist": [
     "last 2 versions",
@@ -48,7 +44,7 @@
     "source/sass/**/*.scss"
   ],
   "engines": {
-    "node": "^16"
+    "node": "16.13.0"
   },
   "type": "module"
 }
