+++ source/catalog.html
@@ -87,11 +87,11 @@
                 <a href="#" class="card__link card__link--img">
                   <picture>
                     <source type="image/webp" media="(min-width:1440px)" srcset="img/chicken-1000-desktop.webp 1x, img/chicken-1000-desktop@2x.webp 2x" width="172" height="196">
-                    <source type="image/webp" media="(min-width:768px)" srcset="img/chicken-1000-tablet.webp 1x, img/chicken-1000-tablet@2x.webp 2x" width="200" height="210">
+                    <source type="image/webp" media="(min-width:768px)" srcset="img/chicken-1000-tablet.webp 1x, img/chicken-1000-tablet@2x.webp 2x" width="196" height="206">
                     <source type="image/webp" srcset="img/chicken-1000-mobile.webp 1x, img/chicken-1000-mobile@2x.webp 2x" width="84" height="100">
 
                     <source media="(min-width:1440px)" srcset="img/chicken-1000-desktop.png 1x, img/chicken-1000-desktop@2x.png 2x" width="172" height="196">
-                    <source media="(min-width:768px)" srcset="img/chicken-1000-tablet.png 1x, img/chicken-1000-tablet@2x.png 2x" width="200" height="210">
+                    <source media="(min-width:768px)" srcset="img/chicken-1000-tablet.png 1x, img/chicken-1000-tablet@2x.png 2x" width="196" height="206">
                     <img class="card__img" src="img/chicken-1000-mobile.png" srcset="img/chicken-1000-mobile@2x.png 2x" width="84" height="100"
                       alt="Cat Energy Pro Chicken 1000.">
                   </picture>
@@ -202,7 +202,7 @@
                 </a>
                 <div class="card__content">
                   <a href="#" class="card__link">
-                    <h2 class="card__title">Cat Energy Pro 500 г</h2>
+                    <h2 class="card__title">Cat Energy Slim 500 г</h2>
                   </a>
                   <ul class="card__properties">
                     <li class="card__property">
@@ -238,7 +238,7 @@
                 </a>
                 <div class="card__content">
                   <a href="#" class="card__link">
-                    <h2 class="card__title">Cat Energy Pro 1000 г</h2>
+                    <h2 class="card__title">Cat Energy Slim 1000 г</h2>
                   </a>
                   <ul class="card__properties">
                     <li class="card__property">
@@ -262,7 +262,7 @@
               <div class="card__wrapper">
                 <a href="#" class="card__link card__link--img">
                   <picture>
-                    <source type="image/webp" media="(min-width:768px)" srcset="img/rice-500.webp 1x, img/rice-500@2x.webp" width="163" height="174">
+                    <source type="image/webp" media="(min-width:768px)" srcset="img/rice-500.webp 1x, img/rice-500@2x.webp 2x" width="163" height="174">
                     <source type="image/webp" srcset="img/rice-500-mobile.webp 1x, img/rice-500-mobile@2x.webp 2x" width="72" height="84">
 
                     <source media="(min-width:768px)" srcset="img/rice-500.png 1x, img/rice-500@2x.png 2x" width="163" height="174">
@@ -272,7 +272,7 @@
                 </a>
                 <div class="card__content">
                   <a href="#" class="card__link">
-                    <h2 class="card__title">Cat Energy Pro 500 г</h2>
+                    <h2 class="card__title">Cat Energy Slim 500 г</h2>
                   </a>
                   <ul class="card__properties">
                     <li class="card__property">
@@ -309,7 +309,7 @@
           <div class="extra__wrapper">
             <ul class="extra__list extra-list">
               <li class="extra-list__item">
-                <h4 class="extra-list__name">Сахорозаменитель</h4>
+                <h4 class="extra-list__name">Сахарозаменитель</h4>
                 <div class="extra-list__values">
                   <p class="extra-list__weight">1 упаковка (100 г)</p>
                   <p class="extra-list__price">200 ₽</p>
@@ -342,7 +342,7 @@
               </li>
             </ul>
             <div class="extra__promotion promotion">
-              <p>Закажите все
+              <p>Закажите все<br>
                 и получите чехол для
                 кота в подарок!</p>
             </div>
@@ -352,7 +352,7 @@
       <section class="map">
         <div class="map__container container">
           <div class="map__info map__info">
-            <h5 class="map__title">Приглашаем к сотрудничеству дилеров!</h5>
+            <h5 class="map__title">Приглашаем<br> к сотрудничеству дилеров!</h5>
             <div class="map__text">
               <span class="map__address">ул. Большая Конюшенная, д. 19/8</span>
               <span class="map__city">Санкт-Петербург</span>
