+++ source/form.html
@@ -45,7 +45,7 @@
     </header>
     <main class="page__main">
       <section class="form">
-        <div class="form__container container">
+        <div class="form__container container container--form">
           <h1 class="form__title">Подбор программы</h1>
           <p class="form__text">Заполните анкету, и мы подберем программу питания для вашего кота</p>
           <form action="https://echo.htmlacademy.ru/">
@@ -96,7 +96,7 @@
                 </label>
                 <label class="form__label field field--tel">
                   <span class="field__name">Телефон:*</span>
-                  <input class="field__input" type="tel" name="telephone" placeholder="8(960)900-60-90" required>
+                  <input class="field__input" type="tel" name="telephone" placeholder="8 (960) 900-60-90" required>
                   <svg class="field__icon" width="20" height="20">
                     <use xlink:href="img/sprite.svg#phone-icon"></use>
                   </svg>
@@ -117,22 +117,22 @@
                 <span>Дополнительно</span>
               </h2>
               <div class="form__checkbox-wrapper">
-                <label class="form__label choice">
+                <label class="form__label choice choice--sweetener">
                   <input class="choice__input" type="checkbox" name="sweetener" value="Cахарозаменитель" checked>
                   <span class="choice__checkbox"></span>
                   <span class="choice__name">Сахарозаменитель</span>
                 </label>
-                <label class="form__label choice">
+                <label class="form__label choice choice--water">
                   <input class="choice__input" type="checkbox" name="water" value="Питьевая вода">
                   <span class="choice__checkbox"></span>
                   <span class="choice__name">Питьевая вода</span>
                 </label>
-                <label class="form__label choice">
+                <label class="form__label choice choice--milk">
                   <input class="choice__input" type="checkbox" name="milk" value="Молоко">
                   <span class="choice__checkbox"></span>
                   <span class="choice__name">Молоко</span>
                 </label>
-                <label class="form__label choice">
+                <label class="form__label choice choice--vitamins">
                   <input class="choice__input" type="checkbox" name="vitamins" value="Витамины">
                   <span class="choice__checkbox"></span>
                   <span class="choice__name">Витамины</span>
@@ -147,9 +147,9 @@
         </div>
       </section>
       <section class="map">
-        <div class="map__container container">
-          <div class="map__info">
-            <h3 class="map__title">Приглашаем к сотрудничеству дилеров!</h3>
+        <div class="map__container container container--form">
+          <div class="map__info map__info--form">
+            <h3 class="map__title">Приглашаем<br> к сотрудничеству дилеров!</h3>
             <div class="map__text">
               <span class="map__address">ул. Большая Конюшенная, д. 19/8</span>
               <span class="map__city">Санкт-Петербург</span>
