+++ source/sass/about.scss
@@ -4,7 +4,7 @@
   @media (min-width: $tablet-width) {
     padding-bottom: 46px;
     border-bottom: 1px solid #e6e6e6;
-    margin-bottom: 45px;
+    margin-bottom: 46px;
   }
 
   @media (min-width: $desktop-width) {
@@ -20,7 +20,7 @@
   line-height: 40px;
   color: #000000;
   margin-top: 0;
-  margin-bottom: 40px;
+  margin-bottom: 41px;
 
   @media (min-width: $tablet-width) {
     font-size: 60px;
@@ -38,8 +38,12 @@
 
   @media (min-width: $tablet-width) {
     gap: 0;
-    justify-content: space-between;
     counter-reset: myCounter;
+    row-gap: 36px;
+  }
+
+  @media (min-width: $desktop-width) {
+    gap: 0;
   }
 }
 
@@ -108,7 +112,15 @@
       }
 
       &::after {
-        right: 26px;
+        right: 5px;
+      }
+
+      @media (min-width: $tablet-width) {
+        margin-right: 153px;
+      }
+
+      @media (min-width: $desktop-width) {
+        margin-right: 94px;
       }
     }
 
@@ -120,7 +132,11 @@
       }
 
       &::after {
-        right: 22px;
+        right: 0;
+      }
+
+      @media (min-width: $desktop-width) {
+        margin-right: 79px;
       }
     }
 
@@ -131,12 +147,17 @@
         width: 207px;
       }
 
+      @media (min-width: $tablet-width) {
+        margin-right: 137px;
+      }
+
       @media (min-width: $desktop-width) {
         width: 248px;
+        margin-right: 77px;
       }
 
       &::after {
-        right: 41px;
+        right: 0;
       }
     }
 
@@ -148,11 +169,15 @@
       }
 
       @media (min-width: $desktop-width) {
-        width: 250px;
+        width: 246px;
       }
 
       &::after {
-        right: 5px;
+        @media (min-width: $tablet-width) {
+          right: 0;
+        }
+
+        right: -4px;
       }
     }
   }
