+++ source/sass/card.scss
@@ -45,6 +45,11 @@
 
       .card__title {
         width: initial;
+        margin-bottom: 8px;
+      }
+
+      .card__button {
+        margin-top: 20px;
       }
     }
 
@@ -90,7 +95,7 @@
       width: 200px;
       height: 210px;
       position: absolute;
-      top: -56px;
+      top: -70px;
       right: 0;
       left: 0;
       margin: 0 auto;
@@ -116,7 +121,7 @@
     font-size: 20px;
     line-height: 24px;
     text-align: center;
-    margin-bottom: 20px;
+    margin-bottom: 21px;
     width: 135px;
   }
 
@@ -134,7 +139,11 @@
   }
 
   @media (min-width: $tablet-width) {
-    margin-bottom: 32px;
+    margin-bottom: 29px;
+  }
+
+  @media (min-width: $desktop-width) {
+    margin-bottom: 37px;
   }
 }
 
@@ -152,7 +161,7 @@
     font-size: 16px;
     line-height: 20px;
     border-bottom: 1px solid #cdcdcd;
-    padding-bottom: 5px;
+    padding-bottom: 6px;
     margin-bottom: 5px;
     width: 250px;
 
@@ -182,7 +191,13 @@
   @media (min-width: $tablet-width) {
     font-size: 16px;
     line-height: 18px;
+    margin-bottom: 82px;
+    width: 170px;
+  }
+
+  @media (min-width: $desktop-width) {
     margin-bottom: 90px;
+    width: initial;
   }
 }
 
@@ -218,6 +233,10 @@
     height: 60px;
     -webkit-mask-size: 60px;
     mask-size: 60px;
+    margin-bottom: 50px;
+  }
+
+  @media (min-width: $desktop-width) {
     margin-bottom: 40px;
   }
 }
