+++ source/sass/catalog.scss
@@ -1,8 +1,12 @@
 .catalog {
-  margin-bottom: 26px;
+  margin-bottom: 25px;
 
   @media (min-width: $tablet-width) {
     margin: 0 auto;
+    margin-bottom: 68px;
+  }
+
+  @media (min-width: $desktop-width) {
     margin-bottom: 76px;
   }
 }
@@ -13,16 +17,17 @@
   line-height: 36px;
   color: #000000;
   margin-top: 0;
-  margin-bottom: 40px;
+  margin-bottom: 42px;
 
   @media (min-width: $tablet-width) {
     max-width: 708px;
     font-size: 60px;
     line-height: 60px;
-    margin-bottom: 165px;
+    margin-bottom: 166px;
   }
 
   @media (min-width: $desktop-width) {
+    margin-bottom: 165px;
     max-width: 1220px;
   }
 }
