+++ source/sass/container.scss
@@ -9,4 +9,10 @@
   @media (min-width: $desktop-width) {
     width: 1220px;
   }
+
+  &--form {
+    @media (min-width: $tablet-width) and (max-width: $desktop-width) {
+      width: 628px;
+    }
+  }
 }
