+++ source/sass/extra-list.scss
@@ -47,6 +47,10 @@
     width: 200px;
     margin: 0;
   }
+
+  @media (min-width: $desktop-width) {
+    width: 285px;
+  }
 }
 
 .extra-list__values {
@@ -63,15 +67,21 @@
     min-width: 223px;
     font-size: 16px;
     line-height: 20px;
+    justify-content: end;
   }
 
   @media (min-width: $desktop-width) {
     min-width: 285px;
+    justify-content: space-between;
   }
 }
 
 .extra-list__weight {
   margin: 0;
+
+  @media (min-width: $tablet-width) {
+    margin-right: 53px;
+  }
 }
 
 .extra-list__price {
