+++ source/sass/main-footer.scss
@@ -1,9 +1,13 @@
 .main-footer {
   background-color: #f2f2f2;
-  padding: 40px 0;
+  padding: 38px 0;
 
   @media (min-width: $tablet-width) {
-    padding: 66px 0;
+    padding: 50px 0;
+  }
+
+  @media (min-width: $desktop-width) {
+    padding: 61px 0;
   }
 }
 
@@ -17,12 +21,11 @@
   @media (min-width: $tablet-width) {
     gap: 0;
     flex-direction: row;
-    justify-content: space-between;
   }
 }
 
 .main-footer__logo {
-  padding-bottom: 19px;
+  padding-bottom: 14px;
   border-bottom: 1px solid #d9d9d9;
   width: 280px;
   text-align: center;
@@ -32,6 +35,11 @@
     border: none;
     width: initial;
     text-align: left;
+    margin-right: 161px;
+  }
+
+  @media (min-width: $desktop-width) {
+    margin-right: 416px;
   }
 
   &:hover {
@@ -54,6 +62,11 @@
     padding: 0;
     border: none;
     justify-content: initial;
+    margin-right: 124px;
+  }
+
+  @media (min-width: $desktop-width) {
+    margin-right: 381px;
   }
 }
 
@@ -64,7 +77,6 @@
   line-height: 20px;
   color: #444444;
   text-decoration: none;
-  padding: 7px 0;
   display: flex;
   align-items: center;
   justify-content: space-between;
