+++ source/sass/main-header.scss
@@ -29,8 +29,10 @@
   }
 
   &--nojs {
-    padding: 0;
-    border: none;
+    @media (max-width: $mobile-width-only) {
+      padding: 0;
+      border: none;
+    }
   }
 }
 
@@ -41,6 +43,10 @@
 
   @media (max-width: $mobile-width-only) {
     flex-wrap: wrap;
+  }
+
+  @media (min-width: $tablet-width) {
+    justify-content: initial;
   }
 }
 
@@ -55,10 +61,12 @@
 
   @media (min-width: $tablet-width) {
     height: 50px;
+    margin-right: 95px;
   }
 
   @media (min-width: $desktop-width) {
     height: 59px;
+    margin-right: 540px;
   }
 
   &::after {
