+++ source/sass/map.scss
@@ -19,7 +19,7 @@
   z-index: 2;
 
   @media (min-width: $tablet-width) {
-    padding: 60px 70px;
+    padding: 60px 39px;
     justify-content: space-between;
   }
 
@@ -29,6 +29,12 @@
     padding: 60px 78px 60px 80px;
     width: 407px;
     gap: 87px;
+  }
+
+  &--form {
+    @media (min-width: $tablet-width) and (max-width: $desktop-width) {
+      padding: 60px 0;
+    }
   }
 }
 
@@ -66,12 +72,16 @@
     gap: 19px;
     width: 157px;
   }
+
+  @media (min-width: $tablet-width) and (max-width: $desktop-width) {
+    text-align: right;
+  }
 }
 
 .map__image {
   object-fit: cover;
   width: 100%;
-  height: 100%;
+  height: 362px;
   display: block;
 
   @media (min-width: $tablet-width) {
