+++ source/sass/promotion.scss
@@ -2,13 +2,13 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  justify-content: center;
+  justify-content: end;
   background-color: rgba(104, 183, 56, 0.85);
   position: relative;
 
   @media (min-width: $tablet-width) and (max-width: $desktop-width) {
     flex-direction: row;
-    justify-content: space-between;
+    justify-content: start;
     padding-left: 77px;
     padding-right: 125px;
   }
@@ -22,6 +22,18 @@
     color: #ffffff;
     margin: 0;
     width: 161px;
+
+    @media (max-width: $mobile-width-only) {
+      padding-bottom: 45px;
+    }
+
+    @media (min-width: $tablet-width) and (max-width: $desktop-width) {
+      text-align: left;
+    }
+
+    @media (min-width: $desktop-width) {
+      padding-bottom: 43px;
+    }
   }
 
   &::before {
@@ -33,8 +45,15 @@
     background-repeat: no-repeat;
     margin-bottom: 48px;
 
+    @media (min-width: $tablet-width) and (max-width: $desktop-width) {
+      margin-bottom: 5px;
+      margin-right: 265px;
+    }
+
     @media (min-width: $desktop-width) {
-      margin-bottom: 46px;
+      margin: 0;
+      position: absolute;
+      top: 55px;
     }
   }
